Rename request logger middleware
